+++ CHANGELOG.md
@@ -18,6 +18,7 @@ For earlier changes, please refer instead to the Git commit log.
 * `SeqHandler.SourceKey()` was removed from the API ([#13](https://github.com/desertwitch/slog-seq/pull/13))
 ### Fixed
 * Do not allow single event to poison batch on status code 400 ([#13](https://github.com/desertwitch/slog-seq/pull/13))
+* Do not allow single event to poison batch on JSON encode failure ([#13](https://github.com/desertwitch/slog-seq/pull/13))
 
 ## v0.7.0 - 2026-06-16
 ### Added